backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pymysql==1.1.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
email-validator==2.1.0.post1
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
cryptography==41.0.7
requests
